Refactor the render part with builder pattern.
