+++ pkg/bootstrap/manifests/hub/managedcluster-clusterrole.yaml
