+++ pkg/bootstrap/manifests/klusterlet/bootstrap_secret.yaml
