+++ pkg/bootstrap/manifests/klusterlet/cluster_role.yaml
