+++ pkg/bootstrap/manifests/klusterlet/clusterrole_aggregate.yaml
