+++ pkg/bootstrap/manifests/klusterlet/clusterrole_bootstrap.yaml
