+++ pkg/bootstrap/manifests/klusterlet/image_pull_secret.yaml
