+++ pkg/bootstrap/manifests/klusterlet/klusterlet.yaml
