+++ pkg/bootstrap/manifests/klusterlet/operator.yaml
